+++ conjure-codegen/src/example_types/objects/product/alias_as_map_key_example.rs
@@ -103,7 +103,7 @@ pub struct AliasAsMapKeyExample {
     )]
     #[serde(
         rename = "uuids",
-        serialize_with = "conjure_serde::cbor::serialize_map_keys_as_strings",
+        serialize_with = "conjure_object::private::serialize_map_keys_as_strings",
         skip_serializing_if = "std::collections::BTreeMap::is_empty",
         default
     )]
