+++ conjure-codegen/src/objects.rs
@@ -208,7 +208,7 @@ fn serde_field_attr(ctx: &Context, def: &ObjectDefinition, field: &FieldDefiniti
 
         if matches!(dealiased_key, Type::Primitive(PrimitiveType::Uuid)) {
             // Use the specialized function for direct UUID keys
-            let serialize_fn = "conjure_serde::cbor::serialize_map_keys_as_strings";
+            let serialize_fn = "conjure_object::private::serialize_map_keys_as_strings";
             parts.push(quote!(serialize_with = #serialize_fn));
         }
     }
