+++ conjure-object/src/private.rs
@@ -355,3 +355,33 @@ impl de::Visitor<'_> for UnionTypeFieldVisitor {
         }
     }
 }
+
+/// Generic version that serializes map keys as strings for Java compatibility.
+///
+/// Works with any key type that implements `Display` (including UUID aliases).
+/// Use this with `#[serde(serialize_with = "conjure_object::private::serialize_map_keys_as_strings")]`
+///
+/// # Example
+/// ```ignore
+/// #[derive(Serialize)]
+/// struct MyType {
+///     #[serde(serialize_with = "conjure_object::private::serialize_map_keys_as_strings")]
+///     uuid_alias_map: BTreeMap<UuidAliasExample, String>,
+/// }
+/// ```
+pub fn serialize_map_keys_as_strings<K, V, S>(
+    map: &BTreeMap<K, V>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    K: std::fmt::Display,
+    V: Serialize,
+    S: serde::Serializer,
+{
+    use serde::ser::SerializeMap;
+    let mut ser_map = serializer.serialize_map(Some(map.len()))?;
+    for (key, value) in map {
+        ser_map.serialize_entry(&key.to_string(), value)?;
+    }
+    ser_map.end()
+}
