+++ conjure-test/test.yml
@@ -183,6 +183,15 @@ types:
         union:
           a: double
           b: RecursiveUnion
+      CborMapKeyTests:
+        docs: Test maps with various key types for CBOR Java compatibility
+        fields:
+          stringKeyMap:
+            type: map<string, string>
+            docs: String keys are the standard case
+          uuidKeyMap:
+            type: map<uuid, string>
+            docs: UUID keys get stringified for Java compatibility
     errors:
       SimpleError:
         namespace: Test
